Issue#54: News CRUD Change package naming from plural to singular Moved UTC Timezone to constants Moved setting createDate and updateDate to business layer Set return type to Page in getAll() service method
