+++ home-service/src/main/java/com/softserveinc/ita/homeproject/homeservice/constants/Timezone.java
@@ -1,0 +1,7 @@
+package com.softserveinc.ita.homeproject.homeservice.constants;
+
+import java.time.ZoneOffset;
+
+public interface Timezone {
+    ZoneOffset UTC_TIMEZONE = ZoneOffset.UTC;
+}
