+++ home-service/src/main/java/com/softserveinc/ita/homeproject/homeservice/dto/CreateOrUpdateNewsDto.java
@@ -11,5 +11,5 @@ import java.time.LocalDateTime;
 @SuperBuilder
 public class CreateOrUpdateNewsDto extends BaseNewsDto {
 
-    private LocalDateTime dateTime;
+    private LocalDateTime createOrUpdateTime;
 }
