+++ home-service/src/main/java/com/softserveinc/ita/homeproject/homeservice/exception/ApiException.java
@@ -1,4 +1,4 @@
-package com.softserveinc.ita.homeproject.homeservice.exceptions;
+package com.softserveinc.ita.homeproject.homeservice.exception;
 
 public class ApiException extends RuntimeException {
     public ApiException(String message) {
